Refactor SolutionUserObject (#29639)
