+++ framework/include/auxkernels/SolutionScalarAux.h
@@ -11,7 +11,7 @@
 
 #include "AuxScalarKernel.h"
 
-class SolutionUserObject;
+class SolutionUserObjectBase;
 
 /**
  * AuxScalarKernel for reading a solution from file.
@@ -32,7 +32,7 @@ protected:
   virtual Real computeValue() override;
 
   /// Reference to the SolutionUserObject storing the solution
-  const SolutionUserObject & _solution_object;
+  const SolutionUserObjectBase & _solution_object;
 
   /// The variable name of interest
   std::string _var_name;
