+++ framework/src/functions/SolutionFunction.C
@@ -9,7 +9,7 @@
 
 #include "MooseError.h"
 #include "SolutionFunction.h"
-#include "SolutionUserObject.h"
+#include "SolutionUserObjectBase.h"
 #include "MooseMesh.h"
 
 registerMooseObject("MooseApp", SolutionFunction);
@@ -57,7 +57,7 @@ SolutionFunction::initialSetup()
   // Get a pointer to the SolutionUserObject. A pointer is used because the UserObject is not
   // available during the
   // construction of the function
-  _solution_object_ptr = &getUserObject<SolutionUserObject>("solution");
+  _solution_object_ptr = &getUserObject<SolutionUserObjectBase>("solution");
 
   std::string var_name;
