+++ test/src/postprocessors/TestDiscontinuousValuePP.C
@@ -9,7 +9,7 @@
 
 // MOOSE includes
 #include "TestDiscontinuousValuePP.h"
-#include "SolutionUserObject.h"
+#include "SolutionUserObjectBase.h"
 
 registerMooseObject("MooseTestApp", TestDiscontinuousValuePP);
 
@@ -46,7 +46,7 @@ TestDiscontinuousValuePP::TestDiscontinuousValuePP(const InputParameters & param
 void
 TestDiscontinuousValuePP::initialSetup()
 {
-  _solution_object_ptr = &getUserObject<SolutionUserObject>("solution");
+  _solution_object_ptr = &getUserObject<SolutionUserObjectBase>("solution");
 }
 
 Real
